Move code for indicator, as it's not a button
